refactor: change ID of ES index to avoid null gmcReferenceumber error
<NO-TICKET>

+++ src/main/java/uk/nhs/hee/tis/revalidation/connection/service/ElasticSearchIndexUpdateHelper.java
@@ -26,8 +26,15 @@ public class ElasticSearchIndexUpdateHelper {
       elasticSearchService.saveExceptionViews(getExceptionViews(connectionInfo));
     }
     else {
-      elasticSearchService.removeExceptionViewByGmcNumber(connectionInfo.getGmcReferenceNumber());
-      elasticSearchService.removeExceptionViewByTcsPersonId(connectionInfo.getTcsPersonId());
+      // If the trainee does not have a gmc number
+      if (connectionInfo.getGmcReferenceNumber() != null) {
+        elasticSearchService.removeExceptionViewByGmcNumber(connectionInfo.getGmcReferenceNumber());
+      }
+
+      // If the trainee does not have a person Id from TCS
+      if (connectionInfo.getTcsPersonId() != null) {
+        elasticSearchService.removeExceptionViewByTcsPersonId(connectionInfo.getTcsPersonId());
+      }
     }
   }
